[skip ci] Move util function from conftest to test_util.py, in order to remove dependency of conftest on torch
